Refactor chat flow into agent orchestrator
